requirements.txt
GitPython
matplotlib
PyQt5